requirements.txt
google-generativeai
google-genai
elevenlabs
sounddevice
numpy
scipy
openai-whisper
fastapi
uvicorn
python-dotenv
requests
twilio